+++ project/src/components/comment-form/comment-form.tsx
@@ -1,44 +1,16 @@
-import { ChangeEvent, FormEvent, useEffect, useState } from 'react';
+import { ChangeEvent, FormEvent,  useCallback,  useEffect, useState } from 'react';
 import { connect, ConnectedProps } from 'react-redux';
 import { bindActionCreators } from 'redux';
 
 import { postCommentAction } from '../../store/api-actions';
 import { ThunkAppDispatch } from '../../types/types';
 import { disableByStarAndLength } from '../../utils/util';
-import { STARS } from '../../constants';
+import CommentFormStars from '../comment-form-stars/comment-form-stars';
+import CommentFormTextarea from '../comment-form-textarea/comment-form-textarea';
+
 
 const ERROR_DISPLAY_TIME = 2000;
 
-type RatingStarProps = {
-  star: {score: number, titleName: string},
-  starsCount: number
-  onChange: (evt: ChangeEvent<HTMLInputElement>) => void,
-  disabled: boolean,
-}
-
-function RatingStar({star: {score, titleName}, starsCount, onChange, disabled}: RatingStarProps) {
-
-  /* eslint-disable no-console */
-  console.log('RatingStar');
-
-  const id = `${score}-stars`;
-
-  return (
-    <>
-      <input
-        disabled={disabled}
-        onChange={onChange}
-        checked={score === starsCount}
-        className="form__rating-input visually-hidden" name="rating" value={score} id={id} type="radio"
-      />
-      <label htmlFor={id} className="reviews__rating-label form__rating-label" title={titleName}>
-        <svg className="form__star-image" width="37" height="33">
-          <use xlinkHref="#icon-star"></use>
-        </svg>
-      </label>
-    </>
-  );
-}
 
 const mapDispatchToProps = (dispatch: ThunkAppDispatch) => bindActionCreators({postComment: postCommentAction}, dispatch);
 const connector = connect(null, mapDispatchToProps);
@@ -88,31 +60,18 @@ function CommentForm({hotelId, postComment} : CommentFormProps): JSX.Element {
   };
 
 
+  const handleTextInput = useCallback( (evt: ChangeEvent<HTMLTextAreaElement>) =>  setReview(evt.target.value), [] ) ;
+  const handleStarClick = useCallback( (evt: ChangeEvent<HTMLInputElement>) => setRating(+evt.currentTarget.value), [] );
+
+
   return (
-    <form className="reviews__form form" action="#" method="post"
-      onSubmit={handleSubmit}
-    >
+    <form className="reviews__form form" action="#" method="post" onSubmit={handleSubmit}>
+
       <label className="reviews__label form__label" htmlFor="review">Your review</label>
-      <div className="reviews__rating-form form__rating">
 
-        {STARS.map((star) => (
-          <RatingStar
-            disabled={isFormBlocked}
-            star={star}
-            starsCount={rating}
-            onChange={() => setRating(star.score)}
-            key={star.score}
-          />))}
+      <CommentFormStars onChange={handleStarClick} isFormBlocked={isFormBlocked} rating={rating}/>
 
-      </div>
-
-      <textarea
-        disabled={isFormBlocked}
-        value={review}
-        onChange={(evt: ChangeEvent<HTMLTextAreaElement>) => setReview(evt.target.value)}
-        className="reviews__textarea form__textarea" id="review" name="review" placeholder="Tell how was your stay, what you like and what can be improved"
-      >
-      </textarea>
+      <CommentFormTextarea onChange={handleTextInput} review={review} isFormBlocked={isFormBlocked}/>
 
       {isErrorSanding && <span style={{color: 'red', fontWeight: 'bold'}}>be careful! you broke everything... but we will fix it. Try again later</span>}
