+++ project/src/components/favorite-btn/favorite-btn.tsx
@@ -16,7 +16,7 @@ type FavoriteBtnProps = PropsFromRedux & {isFavorite: boolean, btnSetting: BtnFa
 
 function FavoriteBtn({isFavorite, btnSetting, authorizationStatus, hotelId, changeStatus, nearbyRoomId = 0}: FavoriteBtnProps & {nearbyRoomId: number}): JSX.Element {
   /* eslint-disable no-console */
-  console.log('FavoriteBtn');
+  // console.log('FavoriteBtn');
 
   const history = useHistory();
